requirements.txt
streamlit~=1.29.0
google-generativeai
python dot-env
python-dotenv~=1.0.0
Pillow~=10.0.1